+++ src/_bonsai/utils.h
@@ -33,7 +33,7 @@ extern char _g_asyncmod;
 char *lowercase(char *str);
 struct berval *create_berval(char *value, long int len);
 PyObject *berval2PyObject(struct berval *bval, int keepbytes);
-int PyObject2char_withlength(PyObject *obj, char **output, long int *len);
+int PyObject2char_withlength(PyObject *obj, char **output, Py_ssize_t *len);
 char *PyObject2char(PyObject *obj);
 struct berval **PyList2BervalList(PyObject *list);
 char **PyList2StringList(PyObject *list);
